changed load_room to an normal int, so we can have innkeepers in rooms higher than 32768 also moved load_room to database storage so we don't have to do a pfile conversion for this change.

+++ code/being.h
@@ -1997,7 +1997,7 @@ class TBeing : public TThing {
     void doSave(silentTypeT, const char * = NULL);
     void doZonefile(const sstring &);
     void doLoot(const sstring &);
-    void saveChar(short int);
+    void saveChar(int);
     int doFlee(const char *);
     int doDisarm(sstring, TThing *);
     int dieReturn(const char *, spellNumT, int);
